+++ lib/src/splash/presentation/pages/splash_page.dart
@@ -1,0 +1,77 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_base/src/splash/application/splash_provider.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class SplashPage extends ConsumerStatefulWidget {
+  const SplashPage({super.key});
+
+  @override
+  ConsumerState<SplashPage> createState() => _SplashPageState();
+}
+
+class _SplashPageState extends ConsumerState<SplashPage>
+    with TickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<double> opacity;
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    precacheImage(const AssetImage('assets/images/splash.png'), context);
+  }
+
+  @override
+  void initState() {
+    _controller = AnimationController(
+      duration: const Duration(milliseconds: 1200),
+      vsync: this,
+    );
+    opacity = TweenSequence(<TweenSequenceItem<double>>[
+      TweenSequenceItem(
+        tween: Tween(begin: 1.0, end: 0.0),
+        weight: 50,
+      ),
+      TweenSequenceItem(
+        tween: Tween(begin: 0.0, end: 1.0),
+        weight: 50,
+      ),
+    ]).animate(
+      CurvedAnimation(
+        parent: _controller,
+        curve: const Interval(
+          0.0,
+          1.0,
+          curve: Curves.ease,
+        ),
+      ),
+    );
+    super.initState();
+    _controller.repeat();
+
+    ref.read(splashProvider.notifier).init();
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      body: Center(
+        child: AnimatedBuilder(
+          animation: _controller,
+          builder: (context, child) {
+            return Opacity(
+              opacity: opacity.value,
+              child: Image.asset('assets/images/splash.png'),
+            );
+          },
+        ),
+      ),
+    );
+  }
+}
